# services/core-api/continuum_api_spec.yaml
openapi: 3.0.3
...
info:
  title: Continuum CRM API
  version: '1.0.0'
  description: |
    An AI-native CRM that captures customer interactions automatically and
    proposes the resulting record changes for human confirmation.

    ## Authentication

    All endpoints except signup, login, provider OAuth callbacks and provider
    webhooks require a bearer token:

        Authorization: Bearer <token>

    Two token types are accepted:

    - **Session JWT** — issued by `/auth/login`, used by the web application.
    - **API token** — prefixed `ctm_`, created in Settings. Subject to the rate
      limit below, and revocable with effect on the very next request.

    ## Rate limiting

    API tokens are limited to 300 requests per minute (FR-API-01). Every
    response carries `X-RateLimit-Limit`, `X-RateLimit-Remaining` and
    `X-RateLimit-Reset`. Requests below the limit are never throttled.

    ## Provenance

    Records that the agent can infer carry a `source` field
    (`HUMAN` | `AGENT_INFERRED` | `IMPORTED` | `ENRICHED`). Deals additionally
    carry `stageSource`, recording whether the current stage was set by a
    person or by a confirmed agent proposal. Consuming clients should surface
    this distinction rather than flattening it.

  contact:
    name: Continuum engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.continuum.example/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /companies:
    get:
      tags: [Companies]
      summary: List companies
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Query' }
      responses:
        '200':
          description: Companies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Company' }
    post:
      tags: [Companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                domain: { type: string, nullable: true }
                website: { type: string, nullable: true }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Company' }
...
  /calls/consent:
    get:
      tags: [Calls]
      summary: Consent regime for a phone number
      description: |
        Resolved from country prefix and area code (FR-PIPE-04). Numbers that
        cannot be classified are treated as all-party: recording someone who did
        not consent is the failure mode worth designing against.
      parameters:
        - name: number
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Consent information
          content:
            application/json:
              schema:
                type: object
                properties:
                  regime: { type: string, enum: [ONE_PARTY, TWO_PARTY, UNKNOWN] }
                  region: { type: string, nullable: true }
                  requiresPrompt: { type: boolean }
                  message: { type: string }
  /calls:
    post:
      tags: [Calls]
      summary: Place a call
      description: |
        Dials through the integrated provider without leaving the application
        (FR-PIPE-03). A call that did not connect is logged as an attempt, never
        as a conversation, so the pipeline does not read healthier than it is.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contactId]
              properties:
                contactId: { type: string, format: uuid }
                dealId: { type: string, format: uuid, nullable: true }
                toNumber: { type: string }
      responses:
        '201':
          description: Call placed
          content:
            application/json:
              schema:
                type: object
                properties:
                  activity: { $ref: '#/components/schemas/Activity' }
                  outcome:
                    type: string
                    enum: [CONNECTED, ATTEMPTED_NO_CONNECT, VOICEMAIL, FAILED]
                  consentRegime: { type: string }
                  consentPromptPlayed: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A session JWT, or a `ctm_`-prefixed API token.
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Query:
      name: q
      in: query
      schema: { type: string }
...
  schemas:
...
    Paginated:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    RecordSource:
      type: string
      description: |
        Where a value came from. Clients should surface this distinction — an
        agent-inferred value must never be presented as indistinguishable from
        one a person typed (FR-DATA-02).
      enum: [HUMAN, AGENT_INFERRED, IMPORTED, ENRICHED]
...
    Company:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        domain: { type: string, nullable: true }
        source: { $ref: '#/components/schemas/RecordSource' }
        tags: { type: array, items: { type: string } }
...
    Activity:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [EMAIL, CALL, MEETING, NOTE] }
        direction: { type: string, enum: [INBOUND, OUTBOUND, INTERNAL] }
        subject: { type: string, nullable: true }
        body:
          type: string
          nullable: true
          description: |
            An excerpt, not the full message. Continuum stores enough to audit
            the summary against; the complete text stays in the user's mailbox
            (data minimization, §30).
        occurredAt: { type: string, format: date-time }
        aiSummary:
          type: string
          nullable: true
          description: Null whenever summarization has not succeeded — never fabricated.
        sentiment:
          type: string
          nullable: true
          enum: [POSITIVE, NEUTRAL, NEGATIVE]
        nextStep:
          type: string
          nullable: true
          description: |
            Explicitly nullable. "No next step" is representable, so the model is
            never forced to invent one (FR-AC-03).
        summaryStatus: { type: string, enum: [PENDING, DONE, FAILED, SKIPPED] }
        callOutcome:
          type: string
          nullable: true
          enum: [CONNECTED, ATTEMPTED_NO_CONNECT, VOICEMAIL, FAILED]
        source: { $ref: '#/components/schemas/RecordSource' }
...